Tila 1535 - Dependencies bump (#497)
* Updates ui app dependancies and some related files

* Updates ui app i18n config to work correctly with next-i18n

* Switches from using react-i19n in next app

* Adds yarn format commands for use with prettier

* Root changes

* Update packages via major. Update some related stuff as well.

* Related changes to version bump

* Related changes to version bump : 2

* OIDC changes

* Common version bump. updates related

* Related changes to version bump : 3

* Updates yarn lock

* Moves util tests to correct project

* Updates tests to reflect changes in jest

* Updates dateselectormenu snapshot

* Related changes to version bump: 4, jest config updated to use SWC instead of babel

* Updates cypress config and renames necessary folders / files

* Related changes to version bump: 5

* Updates based on ts-check errors

* Updates yarn lock

* Updates yarn lock

* Updates calendar typings to get rid of last of the build errors. some smaller fixes as well.

* TILA-1535 cleanup

* TILA-1535 debounce calendar control value change

* FINAL yarn lock update

* Update snapshot...again

* Updates related to version bump: 6

* Version bump required configuration file changes to admin-ui.

* Version bump required file changes to admin-ui.

* Updated conf for i18n translation return types for /ui

* Reverts one test back to original form and makes tests ignored in ts-check

* Updated yarn lock

* cleanup

* cleanup

* Fixing common tests

* getSlotPropGetter

* Adds mapbox-gl to dependencies for dockerfile

* Cypress testing fixes. pt1

* Adding cypress conf to dockerfile

* Adding cypress conf to dockerfile.now correctly named.

* Revert on some changes

* Testing cyppress fixes

* cleanup

* Fixes for translations, map component marker and navigation components, searchurl parsing

* Remove unnecessary cypress in dependencies. already in devdependencies

* DISABLE_AUTH

* Fixing and tweaking related to cypress

* Cypress tests and handlers

* Updated yarn.lock

* 2/8 cypress tests

* cypress reservations fixing wip

* cypress: reservations

* Cypress: All tests pass

* cleanup

* Fixing tests for dockerfile

* Fixes after some weirdness with merging

* Remove some unnecessary logging

* Rebased with main

* Removed unnecessary 'as string' conversions from translation elements. There is a type declaratiion for this in the project root

* Nested hyperlink removed

* Added back missing name for eventsummary

* Fixes after merge

* TILA-2311 fix

* Some carousel fixes

* Removed a not needed 'back' button. Fixed accordion not to have 'close' buttons on the bottom

* Resolving PR comments

* Plural localization fixes

* Breadcrumb and translation fixes

* Missing translation added

* Fixes background color of application round cards

* More translationkey fixes

* Some missing style fixes

* Type for search params. Would be smart to create a proper type for this

* Added missing conditional

* Fixing localizations

* Resolving PR comments

* Resolving last of PR comments

* fixing type-error in axiosclient

* Some formatting errors fixed

---------

+++ admin-ui/src/component/recurring-reservations/Recommendation.tsx
@@ -1,6 +1,6 @@
 import React, { ReactNode, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { useHistory, useParams } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 import styled from "styled-components";
 import trim from "lodash/trim";
 import {
@@ -45,6 +45,7 @@ import { applicantName } from "../applications/util";
 import { useNotification } from "../../context/NotificationContext";
 
 interface IRouteParams {
+  [key: string]: string;
   applicationRoundId: string;
   applicationEventScheduleId: string;
 }
@@ -198,7 +199,7 @@ function Recommendation(): JSX.Element {
   ] = useState<boolean>(false);
 
   const { t } = useTranslation();
-  const history = useHistory();
+  const history = useNavigate();
 
   const { applicationRoundId, applicationEventScheduleId } =
     useParams<IRouteParams>();
@@ -263,7 +264,9 @@ function Recommendation(): JSX.Element {
     } catch (error) {
       notifyError(t("errors.errorSavingRecommendation"));
     } finally {
-      history.push(applicationRoundUrl(applicationRoundId));
+      if (applicationRoundId) {
+        history(applicationRoundUrl(applicationRoundId));
+      }
     }
   };
 
@@ -296,9 +299,9 @@ function Recommendation(): JSX.Element {
         payload
       );
 
-      if (revert) {
+      if (revert && applicationRoundId) {
         await deleteAllocationResult(rec.applicationEventScheduleId);
-        history.push(applicationRoundUrl(applicationRoundId));
+        history(applicationRoundUrl(applicationRoundId));
       }
 
       setActionNotification(revert ? null : "ignored");
@@ -495,28 +498,30 @@ function Recommendation(): JSX.Element {
       />
       {application && recommendation && (
         <>
-          <IngressContainer>
-            <Top>
-              <div>
-                <LinkToOthers
-                  to={`${applicationRoundUrl(applicationRoundId)}/applicant/${
-                    recommendation.applicantId
-                  }`}
-                >
-                  {t("Recommendation.linkToOtherRecommendations")}
-                </LinkToOthers>
-                <Heading>{recommendation.applicationEvent.name}</Heading>
-                <div>{applicationRound?.name}</div>
-                <StyledApplicationEventStatusBlock
-                  status={recommendation.applicationEvent.status}
-                  accepted={recommendation.accepted}
-                />
-              </div>
-              <div>
-                {application && <ApplicantBox application={application} />}
-              </div>
-            </Top>
-          </IngressContainer>
+          {applicationRoundId ? (
+            <IngressContainer>
+              <Top>
+                <div>
+                  <LinkToOthers
+                    to={`${applicationRoundUrl(applicationRoundId)}/applicant/${
+                      recommendation.applicantId
+                    }`}
+                  >
+                    {t("Recommendation.linkToOtherRecommendations")}
+                  </LinkToOthers>
+                  <Heading>{recommendation.applicationEvent.name}</Heading>
+                  <div>{applicationRound?.name}</div>
+                  <StyledApplicationEventStatusBlock
+                    status={recommendation.applicationEvent.status}
+                    accepted={recommendation.accepted}
+                  />
+                </div>
+                <div>
+                  {application && <ApplicantBox application={application} />}
+                </div>
+              </Top>
+            </IngressContainer>
+          ) : null}
           <NarrowContainer>
             {actionNotification === "accepted" && (
               <StyledNotification
